Load time series WRF model run data

Loads time series WRF model run data as a raster package RasterBrick. If the specified data does not exist on the user's machine, then it will try to be downloaded from the AirFire database.

On 2020-09-17, available model identifiers include the following:

PNW-1.33km

PNW-4km

wrf_loadTimeSeries(
  modelName = NULL,
  modelRun = NULL,
  modelRunHours = NULL,
  baseUrl = "http://m2.airfire.org",
  varName = NULL,
  res = NULL,
  xlim = NULL,
  ylim = NULL
)

Arguments

modelName
Model identifier.

modelRun
Model initialization timestamp as "YYYYMMDDHH".

modelRunHours
Vector of hours forecasted from init time, i.e. 7:24.

baseUrl
Base URL for WRF output.

varName
WRF variable to load.

res
Resolution of raster in degrees.

xlim
A vector of coordinate longitude bounds.

ylim
A vector of coordinate latitude bounds.

Value

A raster package RasterBrick object.

Examples

# \donttest{
# library(WRFmet)
# setWRFDataDir("~/Data/WRF")

#tempTimeSeries <- wrf_loadTimeSeries(
# modelName = "PNW-1.33km",
# modelRun = "2020091712",
# modelRunHours = 7:10,
# varName = "TSK",
# res = 0.05,
#  xlim = c(-125, -117),
# ylim = c(45, 49)
# )

# rasterVis::levelplot(tempTimeSeries)
# }